requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.28.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Web scraping
selenium>=4.0.0
wikipedia>=1.4.0
beautifulsoup4>=4.11.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Progress tracking
tqdm>=4.64.0

# JSON handling
ujson>=5.6.0

# Optional: Jupyter support
jupyter>=1.0.0
ipywidgets>=8.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
